. Added collision checking. Also added temporary movement to the ball, until I start working on sensor integration.

+++ lab3/app/src/main/java/com/example/hifoz/lab3/Ball.java
@@ -2,14 +2,16 @@ package com.example.hifoz.lab3;
 
 import android.graphics.Paint;
 
-
+/**
+ * Data structure containg data on a ball
+ */
 public class Ball{
-    public float x;
-    public float y;
-    public float radius;
+    public int x;
+    public int y;
+    public int radius;
     public Paint paint;
 
-    public Ball(float x, float y, float radius, Paint paint){
+    public Ball(int x, int y, int radius, Paint paint){
         this.x = x;
         this.y = y;
         this.radius = radius;
